+++ test/e2e/traces_mtls_about_to_expire_cert_test.go
@@ -4,6 +4,7 @@ package e2e
 
 import (
 	"fmt"
+	"github.com/kyma-project/telemetry-manager/internal/testutils"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
@@ -14,7 +15,6 @@ import (
 	kitk8s "github.com/kyma-project/telemetry-manager/test/testkit/k8s"
 	"github.com/kyma-project/telemetry-manager/test/testkit/mocks/backend"
 	"github.com/kyma-project/telemetry-manager/test/testkit/mocks/telemetrygen"
-	"github.com/kyma-project/telemetry-manager/test/testkit/tlsgen"
 	"github.com/kyma-project/telemetry-manager/test/testkit/verifiers"
 )
 
@@ -35,7 +35,7 @@ var _ = Describe("Traces mTLS with certificates expiring within 2 weeks", Label(
 			kitk8s.NewNamespace(telemetrygenNs).K8sObject(),
 		)
 
-		serverCerts, clientCerts, err := tlsgen.NewCertBuilder(mockBackendName, mockNs).
+		serverCerts, clientCerts, err := testutils.NewCertBuilder(mockBackendName, mockNs).
 			WithAboutToExpireClientCert().
 			Build()
 		Expect(err).ToNot(HaveOccurred())
