+++ test/testkit/k8s/trace_pipeline_v1beta1.go
@@ -3,12 +3,12 @@ package k8s
 
 import (
 	"fmt"
+	"github.com/kyma-project/telemetry-manager/internal/testutils"
 
 	"github.com/google/uuid"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	telemetryv1beta1 "github.com/kyma-project/telemetry-manager/apis/telemetry/v1beta1"
-	"github.com/kyma-project/telemetry-manager/test/testkit/tlsgen"
 )
 
 type tracePipelineV1Beta1 struct {
@@ -41,7 +41,7 @@ func (p *tracePipelineV1Beta1) WithOutputEndpointFromSecret(otlpEndpointRef *tel
 	return p
 }
 
-func (p *tracePipelineV1Beta1) WithTLS(certs tlsgen.ClientCerts) *tracePipelineV1Beta1 {
+func (p *tracePipelineV1Beta1) WithTLS(certs testutils.ClientCerts) *tracePipelineV1Beta1 {
 	p.tls = &telemetryv1beta1.OTLPTLS{
 		Insecure:           false,
 		InsecureSkipVerify: false,
