+++ test/testkit/mocks/backend/backend.go
@@ -2,6 +2,7 @@ package backend
 
 import (
 	"fmt"
+	"github.com/kyma-project/telemetry-manager/internal/testutils"
 	"path/filepath"
 	"strconv"
 
@@ -12,7 +13,6 @@ import (
 	"github.com/kyma-project/telemetry-manager/test/testkit/apiserverproxy"
 	kitk8s "github.com/kyma-project/telemetry-manager/test/testkit/k8s"
 	"github.com/kyma-project/telemetry-manager/test/testkit/mocks/backend/fluentd"
-	"github.com/kyma-project/telemetry-manager/test/testkit/tlsgen"
 )
 
 type SignalType string
@@ -35,7 +35,7 @@ type Backend struct {
 	signalType SignalType
 
 	persistentHostSecret bool
-	certs                *tlsgen.ServerCerts
+	certs                *testutils.ServerCerts
 
 	ConfigMap        *ConfigMap
 	FluentDConfigMap *fluentd.ConfigMap
@@ -60,7 +60,7 @@ func New(name, namespace string, signalType SignalType, opts ...Option) *Backend
 	return backend
 }
 
-func WithTLS(certKey tlsgen.ServerCerts) Option {
+func WithTLS(certKey testutils.ServerCerts) Option {
 	return func(b *Backend) {
 		b.certs = &certKey
 	}
